refactor: move explorer widget into explorer file
