+++ src/widgets/explorer/explorer.rs
@@ -5,12 +5,14 @@ use std::rc::Rc;
 use derivative::Derivative;
 use lotus_lib::cache_pair::CachePairReader;
 use lotus_lib::toc::{DirectoryNode, Node, NodeKind};
-use ratatui::widgets::WidgetRef;
+use ratatui::buffer::Buffer;
+use ratatui::layout::Rect;
+use ratatui::text::{Span, Text};
+use ratatui::widgets::{List, ListState, WidgetRef};
 
 use crate::input::KeyInput;
 
 use super::theme::Theme;
-use super::widget::Renderer;
 
 #[derive(Clone, Derivative)]
 #[derivative(Debug, PartialEq, Eq, Hash)]
@@ -37,11 +39,6 @@ impl<'a> Explorer<'a> {
         file_explorer.get_and_set_files()?;
 
         Ok(file_explorer)
-    }
-
-    #[inline]
-    pub const fn widget(&self) -> impl WidgetRef + '_ {
-        Renderer(self)
     }
 
     pub fn handle<I: Into<KeyInput>>(&mut self, input: I) -> Result<()> {
@@ -151,5 +148,71 @@ impl<'a> Explorer<'a> {
         }
 
         Ok(())
+    }
+}
+
+impl WidgetRef for Explorer<'_> {
+    fn render_ref(&self, area: Rect, buf: &mut Buffer) {
+        let mut state = ListState::default().with_selected(Some(self.selected_idx()));
+
+        let highlight_style = if self.current().kind() == NodeKind::Directory {
+            self.theme().highlight_dir_style
+        } else {
+            self.theme().highlight_item_style
+        };
+
+        let mut list = List::new(
+            self.files()
+                .iter()
+                .map(|file| file.text(self.cwd(), self.theme())),
+        )
+        .style(self.theme().style)
+        .highlight_spacing(self.theme().highlight_spacing.clone())
+        .highlight_style(highlight_style);
+
+        if let Some(symbol) = self.theme().highlight_symbol.as_deref() {
+            list = list.highlight_symbol(symbol);
+        }
+
+        if let Some(block) = self.theme().block.as_ref() {
+            let mut block = block.clone();
+
+            for title_top in self.theme().title_top(self) {
+                block = block.title_top(title_top)
+            }
+
+            list = list.block(block);
+        }
+
+        ratatui::widgets::StatefulWidgetRef::render_ref(&list, area, buf, &mut state)
+    }
+}
+
+trait NodeExt {
+    fn text(&self, cwd: &PathBuf, theme: &Theme) -> Text<'_>;
+}
+
+impl NodeExt for Node {
+    #[inline]
+    fn text(&self, cwd: &PathBuf, theme: &Theme) -> Text<'_> {
+        let mut name = self.name();
+        let root_path = PathBuf::from("");
+        let parent_name = cwd
+            .parent()
+            .unwrap_or(&root_path)
+            .file_name()
+            .and_then(|name| name.to_str())
+            .unwrap_or("");
+        if name.as_str() == parent_name {
+            name = "..".to_string();
+        } else if self.kind() == NodeKind::Directory {
+            name.push('/');
+        }
+        let style = if self.kind() == NodeKind::Directory {
+            *theme.dir_style()
+        } else {
+            *theme.item_style()
+        };
+        Span::styled(name, style).into()
     }
 }
