+++ src/widgets/explorer/mod.rs
@@ -1,5 +1,4 @@
 mod explorer;
 mod theme;
-mod widget;
 
 pub use explorer::Explorer;
